Move cmd/frontend/internal/inventory to internal/inventory
This is used by search filters, which cannot be moved out of
`graphqlbackend` if this is in `cmd/frontend/internal`.
