+++ cmd/frontend/graphqlbackend/language_statistics.go
@@ -1,6 +1,6 @@
 package graphqlbackend
 
-import "github.com/sourcegraph/sourcegraph/cmd/frontend/internal/inventory"
+import "github.com/sourcegraph/sourcegraph/internal/inventory"
 
 type languageStatisticsResolver struct {
 	l inventory.Lang
