+++ internal/inventory/context.go
