+++ internal/inventory/entries.go
