+++ internal/inventory/entries_test.go
