+++ internal/inventory/inventory.go
