+++ internal/inventory/inventory_test.go
